INC-37 feat: add the logic for sending the created post to the server (use the api)
