+++ src/features/posts/model/types.ts
