+++ src/features/posts/ui/create-post/crop/cropPanel/CropPanel.module.scss
