+++ src/features/posts/ui/create-post/crop/uploadPanel/UploadPanel.module.scss
