+++ src/features/posts/ui/create-post/details/info/InfoPost.module.scss
